Add `redis_cmd!` macro to compose Redis commands
