+++ actix-redis/src/error.rs
@@ -1,0 +1,14 @@
+use derive_more::{Display, Error, From};
+
+/// General purpose actix redis error
+#[derive(Debug, Display, Error, From)]
+pub enum Error {
+    #[display(fmt = "Redis error {}", _0)]
+    Redis(redis_async::error::Error),
+    /// Receiving message during reconnecting
+    #[display(fmt = "Redis: Not connected")]
+    NotConnected,
+    /// Cancel all waters when connection get dropped
+    #[display(fmt = "Redis: Disconnected")]
+    Disconnected,
+}
